+++ CHANGELOG.md
@@ -13,6 +13,7 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 * [mse240710: Was the fringe always there on the up-rep icon?](http://meta.stackexchange.com/q/240710) (CSS)
 * [mse240790: Welcome back, user. Click here to get error 404](http://meta.stackoverflow.com/q/240790)
 * [mse243519: Dangling signature dash in comments](http://meta.stackoverflow.com/q/243519)
+* [mse84296: RTL text can mess up comment timestamps](http://meta.stackexchange.com/q/84296) (CSS)
 
 **Removed fixes:**